scripts/repo_parser/requirements.txt
langgraph
langchain-core
langchain-community

langchain-google-genai      
openai                       

requests
python-dotenv

nbformat                    
pyyaml                      
toml                         
markdown-it-py              

regex
tqdm
            
